+++ test/integrationTest/testAssets/testRoot/Lib.cc
@@ -1,0 +1,8 @@
+#include <iostream>
+
+#include "Lib.h"
+
+void printer()
+{
+    std::cout << "Hello World!!!" << std::endl;
+}